refactor: move leases + review packages from internal/ to pkg/
Same Phase 0.1 pattern that moved internal/runtime/proxy/ →
pkg/runtime/proxy/ in #326. The runtime proxy's hooks expose types
from these packages in their public API:

- ToolUseHooks.ReviewCache is review.HeldApprovalCache
- ToolUseHooks.Leases is leases.Service

External consumers (clawvisor-cloud, where the runtime proxy fleet
gets wired into a managed service) need to construct these to install
hooks. With the packages in internal/, the Go internal/ rule blocks
the imports. Moving them to pkg/ unblocks external wiring without any
behavior or API change.

Caller imports updated:
- pkg/clawvisor/run.go
- pkg/runtime/proxy/tooluse_runtime.go and tooluse_runtime_test.go
- pkg/runtime/proxy/session_settings_test.go
- internal/api/handlers/runtime.go and runtime_test.go
- internal/e2e/harness/server.go
- docs/runtime-capability-matrix.md (doc references)

Verified: go vet, go build, go test ./... all pass.
