+++ docs/runtime-capability-matrix.md
@@ -18,13 +18,13 @@ coverage anchors in this repo.
 | OpenAI Responses function-call interception | `internal/runtime/conversation/openai_response.go`, `pkg/runtime/proxy/tooluse_runtime.go` | `api.openai.com /v1/responses` | observe, enforce | approvals, leases | `runtime.tool_use.*`, `runtime.lease.*` | U: `internal/runtime/conversation/*`; I: `pkg/runtime/proxy/tooluse_runtime_test.go`; R: `internal/runtime/conversation/testdata/providers/openai_responses/*`; M: manual pack 6 |
 | OpenAI Chat tool-call interception | `internal/runtime/conversation/openai_response.go` | `api.openai.com /v1/chat/completions` | observe, enforce | approvals | `runtime.tool_use.*`, `runtime.observe.would_prompt_inline` | U: `internal/runtime/conversation/response_test.go`; R: `internal/runtime/conversation/testdata/providers/openai_chat/*`; M: manual pack 6 |
 | Codex responses interception | `internal/runtime/conversation/parser.go`, `openai_response.go` | `chatgpt.com /backend-api/codex/responses` | observe, enforce | approvals, leases | `runtime.tool_use.*`, `runtime.lease.*` | U: `internal/runtime/conversation/parser_test.go`; R: `internal/runtime/conversation/testdata/providers/codex_responses/*`; M: manual pack 6 |
-| Held tool-use approval/release | `pkg/runtime/proxy/tooluse_runtime.go`, `internal/runtime/review/cache.go` | Anthropic, OpenAI function-style, Codex function-style | observe, enforce | `approval_records`, review cache, leases | `runtime.tool_use.held`, `released`, `denied`, `review_failed` | I: `pkg/runtime/proxy/tooluse_runtime_test.go`; M: manual pack 6 |
+| Held tool-use approval/release | `pkg/runtime/proxy/tooluse_runtime.go`, `pkg/runtime/review/cache.go` | Anthropic, OpenAI function-style, Codex function-style | observe, enforce | `approval_records`, review cache, leases | `runtime.tool_use.held`, `released`, `denied`, `review_failed` | I: `pkg/runtime/proxy/tooluse_runtime_test.go`; M: manual pack 6 |
 | One-off egress approval/retry | `pkg/runtime/proxy/policy_hook.go`, `internal/api/handlers/runtime.go` | HTTP egress | observe, enforce | `approval_records`, `one_off_approvals` | `runtime.egress.review_required`, `one_off_created`, `one_off_consumed` | I: `pkg/runtime/proxy/policy_hook_test.go`, `internal/api/handlers/runtime_test.go`; M: manual pack 7 |
 | Runtime task promotion | `internal/api/handlers/runtime.go`, `runtime_controls.go` | egress approvals, event promotion | enforce | `tasks`, `active_task_sessions` | approval-created and task-promotion lineage | I: `internal/api/handlers/runtime_test.go`, `runtime_controls_test.go`; M: manual pack 7/10 |
 | Runtime context-judge task binding | `internal/runtime/policy/context_judge.go`, `pkg/runtime/proxy/policy_hook.go` | unmatched egress, unmatched tool-use | observe, enforce | approvals or matched task attribution | runtime allow/review events plus audit attribution | U: `internal/runtime/policy/context_judge_test.go`; I: `pkg/runtime/proxy/policy_hook_test.go`; M: manual pack 7 |
 | Global runtime policy allow/deny/review | `internal/runtime/policy/*`, `pkg/runtime/proxy/policy_hook.go`, `tooluse_runtime.go` | egress, tool-use | observe, enforce | `runtime_policy_rules` | `runtime.policy.allow_matched`, `deny_matched`, `review_matched`, `runtime.observe.would_*` | I: proxy scenario suite; M: manual pack 10 |
 | Active task binding start/end | `internal/api/handlers/tasks.go`, `internal/mcp/tools_exec.go` | runtime sessions | n/a | `active_task_sessions` | no dedicated runtime event today | I: `internal/api/handlers/tasks_runtime_start_test.go`, `internal/mcp/tools_test.go`; M: manual pack 7 |
-| Lease open/close/correlation | `internal/runtime/leases`, `tooluse_runtime.go` | held/released tool execution | observe, enforce | `tool_execution_leases` | `runtime.lease.opened`, `runtime.lease.closed` | I: `pkg/runtime/proxy/tooluse_runtime_test.go`; M: manual pack 12 |
+| Lease open/close/correlation | `pkg/runtime/leases`, `tooluse_runtime.go` | held/released tool execution | observe, enforce | `tool_execution_leases` | `runtime.lease.opened`, `runtime.lease.closed` | I: `pkg/runtime/proxy/tooluse_runtime_test.go`; M: manual pack 12 |
 | Inbound secret capture | `pkg/runtime/proxy/inbound_secret_runtime.go` | supported provider request bodies | observe, enforce | `runtime_placeholders` | `runtime.autovault.captured`, `observed`, unsupported-surface event | I: `pkg/runtime/proxy/inbound_secret_runtime_test.go`; M: manual pack 8 |
 | Outbound placeholder swap | `pkg/runtime/proxy/placeholder_runtime.go` | credential-bearing headers | observe, enforce | vault lookups, placeholders | `runtime.autovault.authorized` | I: `pkg/runtime/proxy/placeholder_runtime_test.go`; M: manual pack 8/9 |
 | Outbound raw credential detection | `internal/runtime/autovault/detect.go`, `placeholder_runtime.go` | outbound headers only | observe, strict | none by default | `runtime.autovault.observed` | U: detector helpers; I: `pkg/runtime/proxy/placeholder_runtime_test.go`; M: manual pack 9 |
