+++ internal/api/handlers/runtime_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/clawvisor/clawvisor/internal/api/middleware"
 	runtimeproxy "github.com/clawvisor/clawvisor/pkg/runtime/proxy"
-	runtimereview "github.com/clawvisor/clawvisor/internal/runtime/review"
+	runtimereview "github.com/clawvisor/clawvisor/pkg/runtime/review"
 	"github.com/clawvisor/clawvisor/internal/store/sqlite"
 	intvault "github.com/clawvisor/clawvisor/internal/vault"
 	"github.com/clawvisor/clawvisor/pkg/config"
